+++ src/config/key.rs
@@ -1,5 +1,5 @@
-use crate::event_handler::DISGUISED_EVENT_OFFSETTER;
-use evdev::Key;
+use crate::event_handler::{DISGUISED_EVENT_OFFSETTER, KEY_MATCH_ANY};
+use evdev::KeyCode as Key;
 use serde::{Deserialize, Deserializer};
 use std::error::Error;
 use std::str::FromStr;
@@ -119,6 +119,7 @@ pub fn parse_key(input: &str) -> Result<Key, Box<dyn Error>> {
             REL_WHEEL_HI_RES = 0x0b,
             REL_HWHEEL_HI_RES = 0x0c,
         */
+        "ANY" => KEY_MATCH_ANY,
         // End of custom scancodes
 
         // else